Try to fix unrelated failing acceptance tests
