+++ test/Microsoft.TestPlatform.AcceptanceTests/AcceptanceTestBase.cs
@@ -146,4 +146,44 @@ public class AcceptanceTestBase : IntegrationTestBase
 
         return runSettingsXml;
     }
+
+    protected string GetIsolatedTestAsset(string assetName)
+    {
+        var projectPath = GetProjectFullPath(assetName);
+
+        foreach (var file in new FileInfo(projectPath).Directory!.EnumerateFiles())
+        {
+            var newFilePath = Path.Combine(TempDirectory.Path, file.Name);
+
+            if (file.Extension.Equals(".csproj", StringComparison.OrdinalIgnoreCase))
+            {
+                const string testAssetsProps = "TestAssets.props";
+                const string relativePathToRoot = @"..\..\..";
+                const string relativePathToBuild = relativePathToRoot + @"\scripts\build\";
+                const string testAssetsPropsFullPath = relativePathToBuild + testAssetsProps;
+
+                // Copy csproj and edit path to TestAssets.props
+                var content = File.ReadAllText(file.FullName).Replace(testAssetsPropsFullPath, testAssetsProps);
+                File.WriteAllText(newFilePath, content);
+
+                // Copy TestAssets.props and update path to TestPlatform.Dependencies.props
+                const string tpDependenciesPropsFileName = "TestPlatform.Dependencies.props";
+                var assetPropsContent = File.ReadAllText(Path.Combine(file.DirectoryName!, testAssetsPropsFullPath))
+                    .Replace("$(MSBuildThisFileDirectory)" + tpDependenciesPropsFileName, Path.Combine(file.DirectoryName!, relativePathToBuild, tpDependenciesPropsFileName));
+                File.WriteAllText(Path.Combine(TempDirectory.Path, testAssetsProps), assetPropsContent);
+
+                // Copy nuget.config and make packages folder point to vstest packages folder
+                const string nugetFileName = "NuGet.config";
+                var nugetContent = File.ReadAllText(Path.Combine(file.DirectoryName!, relativePathToRoot, nugetFileName))
+                    .Replace("\"packages\"", "\"" + Path.Combine(file.DirectoryName!, relativePathToRoot, "packages") + "\"");
+                File.WriteAllText(Path.Combine(TempDirectory.Path, nugetFileName), nugetContent);
+            }
+            else
+            {
+                File.Copy(file.FullName, newFilePath);
+            }
+        }
+
+        return Path.Combine(TempDirectory.Path, assetName);
+    }
 }
